+++ src/battle_anim_throw.c
@@ -1780,7 +1780,7 @@ static void PokeBallOpenParticleAnimation_Step2(struct Sprite *sprite)
 
 static void TimerBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1814,7 +1814,7 @@ static void TimerBallOpenParticleAnimation(u8 taskId)
 
 static void DiveBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1849,7 +1849,7 @@ static void DiveBallOpenParticleAnimation(u8 taskId)
 // Also used for Net Ball
 static void SafariBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1884,7 +1884,7 @@ static void SafariBallOpenParticleAnimation(u8 taskId)
 // Also used for Nest Ball
 static void UltraBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1919,7 +1919,7 @@ static void UltraBallOpenParticleAnimation(u8 taskId)
 // Also used for Luxury Ball
 static void GreatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1975,7 +1975,7 @@ static void FanOutBallOpenParticles_Step1(struct Sprite *sprite)
 
 static void RepeatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -2017,7 +2017,7 @@ static void RepeatBallOpenParticleAnimation_Step1(struct Sprite *sprite)
 
 static void MasterBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i, j;
+    u32 i, j;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -2063,7 +2063,7 @@ static void MasterBallOpenParticleAnimation(u8 taskId)
 
 static void ThiefBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
