+++ src/field_effect.c
@@ -876,7 +876,7 @@ void FieldEffectStop(struct Sprite *sprite, u8 id)
 
 void FieldEffectFreeTilesIfUnused(u16 tileStart)
 {
-    u8 i;
+    u32 i;
     u16 tag = GetSpriteTileTagByTileStart(tileStart);
 
     if (tag != 0xFFFF)
@@ -890,7 +890,7 @@ void FieldEffectFreeTilesIfUnused(u16 tileStart)
 
 void FieldEffectFreePaletteIfUnused(u8 paletteNum)
 {
-    u8 i;
+    u32 i;
     u16 tag = GetSpritePaletteTagByPaletteNum(paletteNum);
 
     if (tag != 0xFFFF)
@@ -909,7 +909,7 @@ void FieldEffectActiveListClear(void)
 
 void FieldEffectActiveListAdd(u8 id)
 {
-    u8 i;
+    u32 i;
     for (i = 0; i < ARRAY_COUNT(sActiveList); i++)
     {
         if (sActiveList[i] == 0xFF)
@@ -922,7 +922,7 @@ void FieldEffectActiveListAdd(u8 id)
 
 void FieldEffectActiveListRemove(u8 id)
 {
-    u8 i;
+    u32 i;
     for (i = 0; i < ARRAY_COUNT(sActiveList); i++)
     {
         if (sActiveList[i] == id)
@@ -935,7 +935,7 @@ void FieldEffectActiveListRemove(u8 id)
 
 bool8 FieldEffectActiveListContains(u8 id)
 {
-    u8 i;
+    u32 i;
     for (i = 0; i < ARRAY_COUNT(sActiveList); i++)
         if (sActiveList[i] == id)
             return TRUE;
@@ -1010,7 +1010,7 @@ void FreeResourcesAndDestroySprite(struct Sprite *sprite, u8 spriteId)
 }
 
 // r, g, b are between 0 and 16
-void MultiplyInvertedPaletteRGBComponents(u16 i, u8 r, u8 g, u8 b)
+void MultiplyInvertedPaletteRGBComponents(u32 i, u8 r, u8 g, u8 b)
 {
     int curRed, curGreen, curBlue;
     u16 color = gPlttBufferUnfaded[i];
@@ -1031,7 +1031,7 @@ void MultiplyInvertedPaletteRGBComponents(u16 i, u8 r, u8 g, u8 b)
 }
 
 // r, g, b are between 0 and 16
-void MultiplyPaletteRGBComponents(u16 i, u8 r, u8 g, u8 b)
+void MultiplyPaletteRGBComponents(u32 i, u8 r, u8 g, u8 b)
 {
     int curRed, curGreen, curBlue;
     u16 color = gPlttBufferUnfaded[i];
@@ -2810,7 +2810,7 @@ static void VBlankCB_FieldMoveShowMonOutdoors(void)
 
 static void LoadFieldMoveOutdoorStreaksTilemap(u16 offs)
 {
-    u16 i;
+    u32 i;
     u16 *dest;
     dest = (u16 *)(VRAM + ARRAY_COUNT(sFieldMoveStreaksOutdoors_Tilemap) + offs);
     for (i = 0; i < ARRAY_COUNT(sFieldMoveStreaksOutdoors_Tilemap); i++, dest++)
@@ -2942,7 +2942,7 @@ static void AnimateIndoorShowMonBg(struct Task *task)
 
 static bool8 SlideIndoorBannerOnscreen(struct Task *task)
 {
-    u16 i;
+    u32 i;
     u16 srcOffs;
     u16 dstOffs;
     u16 *dest;
@@ -2971,7 +2971,7 @@ static bool8 SlideIndoorBannerOnscreen(struct Task *task)
 
 static bool8 SlideIndoorBannerOffscreen(struct Task *task)
 {
-    u16 i;
+    u32 i;
     u16 dstOffs;
     u16 *dest;
 
@@ -3194,7 +3194,7 @@ static void SurfToolFieldEffect_CheckHeldMovementStatus(struct Task *task)
 
 u8 FldEff_RayquazaSpotlight(void)
 {
-    u8 i, j, k;
+    u32 i, j, k;
     u8 spriteId = CreateSprite(gFieldEffectObjectTemplatePointers[FLDEFFOBJ_RAYQUAZA], 120, -24, 1);
     struct Sprite *sprite = &gSprites[spriteId];
