+++ src/field_player_avatar.c
@@ -415,7 +415,7 @@ static bool8 TryDoMetatileBehaviorForcedMovement(void)
 
 static u8 GetForcedMovementByMetatileBehavior(void)
 {
-    u8 i;
+    u32 i;
 
     if (!(gPlayerAvatar.flags & PLAYER_AVATAR_FLAG_5))
     {
@@ -816,7 +816,7 @@ static bool8 TryPushBoulder(s16 x, s16 y, u8 direction)
 
 static void CheckAcroBikeCollision(s16 x, s16 y, u8 metatileBehavior, u8 *collision)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < ARRAY_COUNT(sAcroBikeTrickMetatiles); i++)
     {
@@ -1365,7 +1365,7 @@ u8 GetPlayerAvatarGenderByGraphicsId(u16 gfxId)
 
 bool8 PartyHasMonWithHeadbutt(void)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < PARTY_SIZE; i++)
     {
@@ -1379,7 +1379,7 @@ bool8 PartyHasMonWithHeadbutt(void)
 
 bool8 PartyHasMonWithSurf(void)
 {
-    u8 i;
+    u32 i;
 
     if (!TestPlayerAvatarFlags(PLAYER_AVATAR_FLAG_SURFING))
     {
@@ -1396,7 +1396,7 @@ bool8 PartyHasMonWithSurf(void)
 
 bool8 PartyHasMonWithWhirlpool(void)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < PARTY_SIZE; i++)
     {
@@ -1444,7 +1444,7 @@ void SetPlayerAvatarStateMask(u8 flags)
 
 static u8 GetPlayerAvatarStateTransitionByGraphicsId(u16 graphicsId, u8 gender)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < ARRAY_COUNT(sPlayerAvatarGfxToStateFlag[0]); i++)
     {
@@ -1456,7 +1456,7 @@ static u8 GetPlayerAvatarStateTransitionByGraphicsId(u16 graphicsId, u8 gender)
 
 u16 GetPlayerAvatarGraphicsIdByCurrentState(void)
 {
-    u8 i;
+    u32 i;
     u8 flags = gPlayerAvatar.flags;
 
     for (i = 0; i < ARRAY_COUNT(sPlayerAvatarGfxToStateFlag[0]); i++)
